Move concept and note creation out into model/*
